requirements.txt
openai
pyaudio
pvporcupine
SpeechRecognition
pyttsx3
python-dotenv
loguru
requests
apscheduler
pydantic
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
pydantic-settings
sounddevice
flac

